requirements.txt
django
django-simple-history
Pillow
django-debug-toolbar
psycopg2
openpyxl
